Check that variable assignments only have a single incoming edge
WIP still! Just want to make sure the logic is correct before I start writing
tests. Also is there a good Exception for this case? I just threw in RTE for it
to compile but I'm sure there's something better. Maybe something we catch and
use the list of strings to print a helpful error message?

issue #51